[WIP][394]remove:repeat password field
