+++ ios/models/UserModel/UserModel.h
@@ -7,26 +7,19 @@
 
 #import <Foundation/Foundation.h>
 #import "UserModelObservable.h"
+#import "UserModelConstants.h"
 
 NS_ASSUME_NONNULL_BEGIN
-
-typedef NS_ENUM(NSInteger, UserModelState) {
-    UserModelStateSignUpForm,
-    UserModelStateSignUpEmailInProgress,
-    UserModelStateCodeConfirmForm,
-    UserModelStateCodeConfirmInProgress,
-};
 
 @protocol UserModelObserver;
 @class UserState;
 
 @interface UserModel : NSObject<UserModelObservable>
 
-@property (strong, nonatomic) UserState *emailSendingState;
 @property (strong, nonatomic) NSString *email;
 @property (strong, nonatomic) NSString *password;
 @property (strong, nonatomic) NSMutableArray<id<UserModelObserver>> *userModelObservers;
-@property (assign, nonatomic) UserModelState *state;
+@property (assign, nonatomic) UserModelState state;
 
 @end
