+++ ios/models/UserModel/UserModelConstants.h
@@ -1,0 +1,20 @@
+//
+//  UserModelConstants.h
+//  greenTravel
+//
+//  Created by Alex K on 7.06.22.
+//
+
+#ifndef UserModelConstants_h
+#define UserModelConstants_h
+
+
+#endif /* UserModelConstants_h */
+
+typedef NS_ENUM(NSInteger, UserModelState) {
+    UserModelStateSignUpForm,
+    UserModelStateSignUpEmailInProgress,
+    UserModelStateCodeConfirmForm,
+    UserModelStateCodeConfirmInProgress,
+    UserModelStateSignUpSuccess,
+};
